Move the LDAP encryptors to their own files
